Address Safer cpp warnings in NetworkProcess/cache & NetworkProcess/webrtc https://bugs.webkit.org/show_bug.cgi?id=299428
Reviewed by Geoffrey Garen.

Address Safer cpp warnings in NetworkProcess/cache & NetworkProcess/webrtc.
I had to move some code to a .mm file so that static analysis would properly
recognize the need for `adoptNS()`.

* Source/WTF/wtf/WorkQueue.h:
(WTF::WorkQueueBase::protectedDispatchQueue const):
* Source/WebKit/NetworkProcess/cache/NetworkCacheData.h:
* Source/WebKit/NetworkProcess/cache/NetworkCacheDataCocoa.mm:
(WebKit::NetworkCache::Data::protectedDispatchData const):
(WebKit::NetworkCache::Data::subrange const):
(WebKit::NetworkCache::concatenate):
* Source/WebKit/NetworkProcess/cache/NetworkCacheIOChannelCocoa.mm:
(WebKit::NetworkCache::IOChannel::read):
(WebKit::NetworkCache::IOChannel::write):
* Source/WebKit/NetworkProcess/webrtc/NetworkRTCMonitor.cpp:
(WebKit::gatherNetworkMap):
(WebKit::NetworkRTCMonitor::~NetworkRTCMonitor):
(WebKit::NetworkRTCMonitor::ipv4 const):
(WebKit::NetworkRTCMonitor::ipv6 const):
(WebKit::NetworkRTCMonitor::startUpdatingIfNeeded):
(WebKit::NetworkRTCMonitor::stopUpdating):
(WebKit::NetworkRTCSharedMonitor::ipv4 const): Deleted.
(WebKit::NetworkRTCSharedMonitor::ipv6 const): Deleted.
(WebKit::networkSharedMonitor): Deleted.
(WebKit::createNWPathMonitor): Deleted.
(WebKit::NetworkRTCSharedMonitor::NetworkRTCSharedMonitor): Deleted.
(WebKit::NetworkRTCSharedMonitor::addListener): Deleted.
(WebKit::NetworkRTCSharedMonitor::removeListener): Deleted.
(WebKit::interfaceAdapterType): Deleted.
(WebKit::NetworkRTCSharedMonitor::adapterTypeFromInterfaceName const): Deleted.
(WebKit::NetworkRTCSharedMonitor::updateNetworks): Deleted.
(WebKit::NetworkRTCSharedMonitor::updateNetworksFromPath): Deleted.
(WebKit::NetworkRTCSharedMonitor::onGatheredNetworks): Deleted.
* Source/WebKit/NetworkProcess/webrtc/NetworkRTCSharedMonitor.cpp: Added.
(WebKit::NetworkRTCSharedMonitor::singleton):
(WebKit::NetworkRTCSharedMonitor::NetworkRTCSharedMonitor):
(WebKit::NetworkRTCSharedMonitor::addListener):
(WebKit::NetworkRTCSharedMonitor::removeListener):
(WebKit::NetworkRTCSharedMonitor::adapterTypeFromInterfaceName const):
(WebKit::NetworkRTCSharedMonitor::updateNetworks):
(WebKit::NetworkRTCSharedMonitor::onGatheredNetworks):
* Source/WebKit/NetworkProcess/webrtc/NetworkRTCSharedMonitor.h: Added.
(WebKit::NetworkRTCSharedMonitor::ipv4 const):
(WebKit::NetworkRTCSharedMonitor::ipv6 const):
* Source/WebKit/NetworkProcess/webrtc/NetworkRTCSharedMonitorCocoa.mm: Added.
(WebKit::createNWPathMonitor):
(WebKit::interfaceAdapterType):
(WebKit::NetworkRTCSharedMonitor::setupNWPathMonitor):
(WebKit::NetworkRTCSharedMonitor::updateNetworksFromPath):
* Source/WebKit/NetworkProcess/webrtc/NetworkRTCTCPSocketCocoa.mm:
(WebKit::tcpSocketQueueSingleton):
(WebKit::NetworkRTCTCPSocketCocoa::NetworkRTCTCPSocketCocoa):
(WebKit::NetworkRTCTCPSocketCocoa::getInterfaceName):
(WebKit::tcpSocketQueue): Deleted.
* Source/WebKit/NetworkProcess/webrtc/NetworkRTCUDPSocketCocoa.mm:
(WebKit::udpSocketQueueSingleton):
(WebKit::NetworkRTCUDPSocketCocoaConnections::NetworkRTCUDPSocketCocoaConnections):
(WebKit::NetworkRTCUDPSocketCocoaConnections::setupNWConnection):
(WebKit::udpSocketQueue): Deleted.
* Source/WebKit/NetworkProcess/webtransport/cocoa/NetworkTransportSessionCocoa.mm:
* Source/WebKit/SaferCPPExpectations/RetainPtrCtorAdoptCheckerExpectations:
* Source/WebKit/SaferCPPExpectations/UnretainedCallArgsCheckerExpectations:
* Source/WebKit/SourcesCocoa.txt:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/300507@main
