+++ Source/WebKit/NetworkProcess/webrtc/NetworkRTCMonitor.cpp
@@ -31,6 +31,7 @@
 #include "Connection.h"
 #include "Logging.h"
 #include "NetworkRTCProvider.h"
+#include "NetworkRTCSharedMonitor.h"
 #include "WebRTCMonitorMessages.h"
 #include <WebCore/Timer.h>
 #include <ifaddrs.h>
@@ -82,146 +83,12 @@ private:
     RTCNetwork::IPAddress m_ipv6;
 };
 
-class NetworkRTCSharedMonitor {
-public:
-    NetworkRTCSharedMonitor();
-
-    void addListener(NetworkRTCMonitor&);
-    void removeListener(NetworkRTCMonitor&);
-
-    const RTCNetwork::IPAddress& ipv4() const { return m_ipv4; }
-    const RTCNetwork::IPAddress& ipv6()  const { return m_ipv6; }
-
-    webrtc::AdapterType adapterTypeFromInterfaceName(const char*) const;
-
-#if PLATFORM(COCOA)
-    void updateNetworksFromPath(nw_path_t);
-#endif
-
-private:
-    void start();
-    void stop();
-
-    void updateNetworks();
-    void updateNetworksOnQueue();
-
-    void onGatheredNetworks(RTCNetwork::IPAddress&&, RTCNetwork::IPAddress&&, HashMap<String, RTCNetwork>&&);
-
-    WeakHashSet<NetworkRTCMonitor> m_observers;
-
-    Ref<ConcurrentWorkQueue> m_queue;
-    WebCore::Timer m_updateNetworksTimer;
-
-    bool m_didReceiveResults { false };
-    Vector<RTCNetwork> m_networkList;
-    RTCNetwork::IPAddress m_ipv4;
-    RTCNetwork::IPAddress m_ipv6;
-    int m_networkLastIndex { 0 };
-    HashMap<String, RTCNetwork> m_networkMap;
-#if PLATFORM(COCOA)
-    RetainPtr<nw_path_monitor> m_nwMonitor;
-    HashMap<String, webrtc::AdapterType> m_adapterTypes;
-#endif
-};
-
-static NetworkRTCSharedMonitor& networkSharedMonitor()
-{
-    static NeverDestroyed<NetworkRTCSharedMonitor> networkSharedMonitor;
-    return networkSharedMonitor.get();
-}
-
-#if PLATFORM(COCOA)
-static RetainPtr<nw_path_monitor> createNWPathMonitor()
-{
-    auto nwMonitor = adoptCF(nw_path_monitor_create());
-    nw_path_monitor_set_queue(nwMonitor.get(), mainDispatchQueueSingleton());
-    nw_path_monitor_set_update_handler(nwMonitor.get(), makeBlockPtr([](nw_path_t path) {
-        networkSharedMonitor().updateNetworksFromPath(path);
-    }).get());
-    return nwMonitor;
-}
-#endif
-
-NetworkRTCSharedMonitor::NetworkRTCSharedMonitor()
-    : m_queue(ConcurrentWorkQueue::create("NetworkRTCSharedMonitor queue"_s))
-    , m_updateNetworksTimer([] { networkSharedMonitor().updateNetworks(); })
-{
-}
-
-void NetworkRTCSharedMonitor::addListener(NetworkRTCMonitor& monitor)
-{
-    if (m_didReceiveResults)
-        monitor.onNetworksChanged(m_networkList, m_ipv4, m_ipv6);
-
-    bool shouldStart = m_observers.isEmptyIgnoringNullReferences();
-    m_observers.add(monitor);
-
-    RELEASE_LOG(WebRTC, "NetworkRTCSharedMonitor::addListener shouldStart=%d didReceiveResults=%d listener=%p", shouldStart, m_didReceiveResults, &monitor);
-
-    if (!shouldStart)
-        return;
-
-#if PLATFORM(COCOA)
-    if (monitor.rtcProvider().webRTCInterfaceMonitoringViaNWEnabled()) {
-        if (auto nwMonitor = std::exchange(m_nwMonitor, { }))
-            nw_path_monitor_cancel(m_nwMonitor.get());
-
-        RELEASE_LOG(WebRTC, "NetworkRTCSharedMonitor::createNWPathMonitor");
-
-        m_nwMonitor = createNWPathMonitor();
-        nw_path_monitor_start(m_nwMonitor.get());
-        return;
-    }
-#endif
-
-    updateNetworks();
-    m_updateNetworksTimer.startRepeating(2_s);
-}
-
-void NetworkRTCSharedMonitor::removeListener(NetworkRTCMonitor& monitor)
-{
-    m_observers.remove(monitor);
-
-    bool shouldStop = m_observers.isEmptyIgnoringNullReferences();
-
-    RELEASE_LOG(WebRTC, "NetworkRTCSharedMonitor::removeListener shouldStop=%d listener=%p", shouldStop, &monitor);
-
-    if (!shouldStop)
-        return;
-
-#if PLATFORM(COCOA)
-    if (auto nwMonitor = std::exchange(m_nwMonitor, { }))
-        nw_path_monitor_cancel(nwMonitor.get());
-#endif
-
-    m_updateNetworksTimer.stop();
-}
-
 static std::optional<std::pair<RTCNetwork::InterfaceAddress, RTCNetwork::IPAddress>> addressFromInterface(const struct ifaddrs& interface)
 {
     RTCNetwork::IPAddress address { *interface.ifa_addr };
     RTCNetwork::IPAddress mask { *interface.ifa_netmask };
     return std::make_pair(RTCNetwork::InterfaceAddress { address, webrtc::IPV6_ADDRESS_FLAG_NONE }, mask);
 }
-
-#if PLATFORM(COCOA)
-static webrtc::AdapterType interfaceAdapterType(nw_interface_t interface)
-{
-    switch (nw_interface_get_type(interface)) {
-    case nw_interface_type_other:
-        return webrtc::ADAPTER_TYPE_VPN;
-    case nw_interface_type_wifi:
-        return webrtc::ADAPTER_TYPE_WIFI;
-    case nw_interface_type_cellular:
-        return webrtc::ADAPTER_TYPE_CELLULAR;
-    case nw_interface_type_wired:
-        return webrtc::ADAPTER_TYPE_ETHERNET;
-    case nw_interface_type_loopback:
-        return webrtc::ADAPTER_TYPE_LOOPBACK;
-    }
-    return webrtc::ADAPTER_TYPE_UNKNOWN;
-}
-#endif
 
 static HashMap<String, RTCNetwork> gatherNetworkMap()
 {
@@ -258,7 +125,7 @@ static HashMap<String, RTCNetwork> gatherNetworkMap()
         auto networkKey = makeString(name, "-"_s, prefixLength, "-"_s, std::span { prefixString });
 
         networkMap.ensure(networkKey, [&] {
-            auto interfaceType = networkSharedMonitor().adapterTypeFromInterfaceName(iterator->ifa_name);
+            auto interfaceType = NetworkRTCSharedMonitor::singleton().adapterTypeFromInterfaceName(iterator->ifa_name);
             return RTCNetwork { name, networkKey.utf8().span(), address->second, prefixLength, interfaceType, 0, 0, true, false, scopeID, { } };
         }).iterator->value.ips.append(address->first);
     }
@@ -346,54 +213,6 @@ static std::optional<RTCNetwork::IPAddress> getDefaultIPAddress(bool useIPv4)
     return getSocketLocalAddress(socket, useIPv4);
 }
 
-webrtc::AdapterType NetworkRTCSharedMonitor::adapterTypeFromInterfaceName(const char* interfaceName) const
-{
-#if PLATFORM(COCOA)
-    auto iterator = m_adapterTypes.find(String::fromUTF8(interfaceName));
-    if (iterator != m_adapterTypes.end())
-        return iterator->value;
-#endif
-    return webrtc::GetAdapterTypeFromName(interfaceName);
-}
-
-void NetworkRTCSharedMonitor::updateNetworks()
-{
-    auto aggregator = CallbackAggregator::create([] (auto&& ipv4, auto&& ipv6, auto&& networkList) mutable {
-        networkSharedMonitor().onGatheredNetworks(WTFMove(ipv4), WTFMove(ipv6), WTFMove(networkList));
-    });
-    Ref protectedQueue = m_queue;
-    protectedQueue->dispatch([aggregator] {
-        bool useIPv4 = true;
-        if (auto address = getDefaultIPAddress(useIPv4))
-            aggregator->setIPv4(WTFMove(*address));
-    });
-    protectedQueue->dispatch([aggregator] {
-        bool useIPv4 = false;
-        if (auto address = getDefaultIPAddress(useIPv4))
-            aggregator->setIPv6(WTFMove(*address));
-    });
-    protectedQueue->dispatch([aggregator] {
-        aggregator->setNetworkMap(gatherNetworkMap());
-    });
-}
-
-#if PLATFORM(COCOA)
-void NetworkRTCSharedMonitor::updateNetworksFromPath(nw_path_t path)
-{
-    RELEASE_LOG(WebRTC, "NetworkRTCSharedMonitor::updateNetworksFromPath");
-
-    auto status = nw_path_get_status(path);
-    if (status != nw_path_status_satisfied && status != nw_path_status_satisfiable)
-        return;
-
-    nw_path_enumerate_interfaces(path, makeBlockPtr([](nw_interface_t interface) -> bool {
-        networkSharedMonitor().m_adapterTypes.set(String::fromUTF8(nw_interface_get_name(interface)), interfaceAdapterType(interface));
-        return true;
-    }).get());
-    updateNetworks();
-}
-#endif
-
 static bool isEqual(const RTCNetwork::InterfaceAddress& a, const RTCNetwork::InterfaceAddress& b)
 {
     return a.rtcAddress() == b.rtcAddress();
@@ -435,55 +254,6 @@ static bool sortNetworks(const RTCNetwork& a, const RTCNetwork& b)
     return codePointCompare(StringView { a.description.span() }, StringView { b.description.span() }) < 0;
 }
 
-void NetworkRTCSharedMonitor::onGatheredNetworks(RTCNetwork::IPAddress&& ipv4, RTCNetwork::IPAddress&& ipv6, HashMap<String, RTCNetwork>&& networkMap)
-{
-    if (!m_didReceiveResults) {
-        m_didReceiveResults = true;
-        m_networkMap = WTFMove(networkMap);
-        m_ipv4 = WTFMove(ipv4);
-        m_ipv6 = WTFMove(ipv6);
-
-        for (auto& network : m_networkMap.values())
-            network.id = ++m_networkLastIndex;
-    } else {
-        bool didChange = networkMap.size() != networkMap.size();
-        for (auto& keyValue : networkMap) {
-            auto iterator = m_networkMap.find(keyValue.key);
-            bool isFound = iterator != m_networkMap.end();
-            keyValue.value.id = isFound ? iterator->value.id : ++m_networkLastIndex;
-            didChange |= !isFound || hasNetworkChanged(keyValue.value, iterator->value);
-        }
-        if (!didChange) {
-            for (auto& keyValue : m_networkMap) {
-                if (!networkMap.contains(keyValue.key)) {
-                    didChange = true;
-                    break;
-                }
-            }
-        }
-        if (!didChange && (ipv4.isUnspecified() || isEqual(ipv4, m_ipv4)) && (ipv6.isUnspecified() || isEqual(ipv6, m_ipv6)))
-            return;
-
-        m_networkMap = WTFMove(networkMap);
-        if (!ipv4.isUnspecified())
-            m_ipv4 = WTFMove(ipv4);
-        if (!ipv6.isUnspecified())
-            m_ipv6 = WTFMove(ipv6);
-    }
-    RELEASE_LOG(WebRTC, "NetworkRTCSharedMonitor::onGatheredNetworks - networks changed");
-
-    auto networkList = copyToVector(m_networkMap.values());
-    std::ranges::sort(networkList, sortNetworks);
-
-    int preference = std::max(127zu, networkList.size());
-    for (auto& network : networkList)
-        network.preference = preference--;
-
-    m_observers.forEach([this](auto& observer) {
-        Ref { observer }->onNetworksChanged(m_networkList, m_ipv4, m_ipv6);
-    });
-}
-
 NetworkRTCMonitor::NetworkRTCMonitor(NetworkRTCProvider& rtcProvider)
     : m_rtcProvider(rtcProvider)
 {
@@ -491,7 +261,7 @@ NetworkRTCMonitor::NetworkRTCMonitor(NetworkRTCProvider& rtcProvider)
 
 NetworkRTCMonitor::~NetworkRTCMonitor()
 {
-    networkSharedMonitor().removeListener(*this);
+    NetworkRTCSharedMonitor::singleton().removeListener(*this);
 }
 
 NetworkRTCProvider& NetworkRTCMonitor::rtcProvider()
@@ -501,12 +271,12 @@ NetworkRTCProvider& NetworkRTCMonitor::rtcProvider()
 
 const RTCNetwork::IPAddress& NetworkRTCMonitor::ipv4() const
 {
-    return networkSharedMonitor().ipv4();
+    return NetworkRTCSharedMonitor::singleton().ipv4();
 }
 
 const RTCNetwork::IPAddress& NetworkRTCMonitor::ipv6()  const
 {
-    return networkSharedMonitor().ipv6();
+    return NetworkRTCSharedMonitor::singleton().ipv6();
 }
 
 void NetworkRTCMonitor::startUpdatingIfNeeded()
@@ -514,7 +284,7 @@ void NetworkRTCMonitor::startUpdatingIfNeeded()
 #if ASSERT_ENABLED
     m_isStarted = true;
 #endif
-    networkSharedMonitor().addListener(*this);
+    NetworkRTCSharedMonitor::singleton().addListener(*this);
 }
 
 void NetworkRTCMonitor::stopUpdating()
@@ -522,7 +292,7 @@ void NetworkRTCMonitor::stopUpdating()
 #if ASSERT_ENABLED
     m_isStarted = false;
 #endif
-    networkSharedMonitor().removeListener(*this);
+    NetworkRTCSharedMonitor::singleton().removeListener(*this);
 }
 
 void NetworkRTCMonitor::onNetworksChanged(const Vector<RTCNetwork>& networkList, const RTCNetwork::IPAddress& ipv4, const RTCNetwork::IPAddress& ipv6)
