+++ Source/WebKit/SaferCPPExpectations/RetainPtrCtorAdoptCheckerExpectations
@@ -1,6 +1,5 @@
 NetworkProcess/cocoa/NetworkSessionCocoa.mm
 NetworkProcess/mac/SecItemShim.cpp
-NetworkProcess/webrtc/NetworkRTCMonitor.cpp
 Platform/cf/ModuleCF.cpp
 Shared/API/Cocoa/WKRemoteObjectCoder.mm
 Shared/API/Cocoa/_WKRemoteObjectRegistry.mm
